Do not use class-transformer and class-validator
There seems to be no way to prevent the user from specifying additional properties, such as id or date of registration, which he should not be able to edit.
In addition, these dependencies seem redundant, since it is not clear where they can be useful in the project, except for registration.